feat: Enhance ticket order processing and pricing logic, auto-generate basic ticket plans for events without plans

+++ app/Models/Performer.php
@@ -104,7 +104,7 @@ final class Performer extends Model
     {
         // This would be calculated based on user's location
         // For now, return a default value
-        return $this->attributes['distance_miles'] ?? 0.0;
+        return (float) ($this->attributes['distance_miles'] ?? 0.0);
     }
 
     // Scopes
